opea-comps/mega-service-new/app/requirements.txt
opea-comps
fastapi<0.116.0
aiofiles<24.0,>=22.0
protobuf==4.21.6
grpcio==1.59.0
grpcio-status==1.59.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
uvicorn[standard]>=0.27.1
